+++ src/ReceivingList/ReceivingListContainer.js
@@ -1,17 +1,11 @@
 import React, {
   useCallback,
-  useState,
 } from 'react';
 import PropTypes from 'prop-types';
-import {
-  withRouter,
-} from 'react-router-dom';
-import ReactRouterPropTypes from 'react-router-prop-types';
-import queryString from 'query-string';
 import { useIntl } from 'react-intl';
 
 import { stripesConnect } from '@folio/stripes/core';
-import { useList } from '@folio/stripes-acq-components';
+import { usePagination } from '@folio/stripes-acq-components';
 
 import {
   titlesResource,
@@ -21,8 +15,8 @@ import {
 } from '../common/resources';
 import ReceivingList from './ReceivingList';
 
+import { useReceiving } from './hooks';
 import {
-  buildTitlesQuery,
   fetchLinesOrders,
   fetchOrderLineLocations,
   fetchTitleOrderLines,
@@ -32,103 +26,54 @@ const RESULT_COUNT_INCREMENT = 30;
 
 const resetData = () => {};
 
-const ReceivingListContainer = ({ mutator, location }) => {
+const ReceivingListContainer = ({ mutator }) => {
   const intl = useIntl();
 
   const invalidReferenceMessage = intl.formatMessage({ id: 'ui-receiving.titles.invalidReference' });
 
-  const [orderLinesMap, setOrderLinesMap] = useState({});
-  const [locationsMap, setLocationsMap] = useState({});
-  const [ordersMap, setOrdersMap] = useState({});
+  const fetchReferences = useCallback(async titlesResponse => {
+    const orderLinesResponse = await fetchTitleOrderLines(mutator.receivingListOrderLines, titlesResponse, {});
+    const locationsResponse = await fetchOrderLineLocations(mutator.receivingListLocations, orderLinesResponse, {});
+    const linesOrdersResponse = await fetchLinesOrders(mutator.lineOrders, orderLinesResponse, {});
 
-  const loadTitles = useCallback((offset) => {
-    return mutator.receivingListTitles.GET({
-      params: {
-        limit: RESULT_COUNT_INCREMENT,
-        offset,
-        query: buildTitlesQuery(queryString.parse(location.search)),
-      },
-    });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location.search]);
+    const locationsMap = locationsResponse.reduce((acc, locationItem) => {
+      acc[locationItem.id] = locationItem;
 
-  const loadTitlesCB = useCallback((setTitles, titlesResponse) => {
-    const orderLinesPromise = fetchTitleOrderLines(
-      mutator.receivingListOrderLines, titlesResponse.titles, orderLinesMap,
-    );
+      return acc;
+    }, {});
 
-    return orderLinesPromise
-      .then((orderLinesResponse) => {
-        const locationsPromise = fetchOrderLineLocations(
-          mutator.receivingListLocations, orderLinesResponse, locationsMap,
-        );
-        const linesOrdersPromise = fetchLinesOrders(mutator.lineOrders, orderLinesResponse, ordersMap);
+    const ordersMap = linesOrdersResponse.reduce((acc, order) => {
+      acc[order.id] = order;
 
-        return Promise.all([orderLinesResponse, locationsPromise, linesOrdersPromise]);
-      })
-      .then(([orderLinesResponse, locationsResponse, ordersResponse]) => {
-        const newLocationsMap = {
-          ...locationsMap,
-          ...locationsResponse.reduce((acc, locationItem) => {
-            acc[locationItem.id] = locationItem;
+      return acc;
+    }, {});
 
-            return acc;
-          }, {}),
-        };
+    const orderLinesMap = orderLinesResponse.reduce((acc, orderLine) => {
+      acc[orderLine.id] = {
+        ...orderLine,
+        locations: orderLine.locations.map(
+          ({ locationId }) => locationsMap[locationId]?.name ?? invalidReferenceMessage,
+        ),
+        orderWorkflow: ordersMap[orderLine.purchaseOrderId]?.workflowStatus,
+      };
 
-        const newOrdersMap = {
-          ...ordersMap,
-          ...ordersResponse.reduce((acc, d) => {
-            acc[d.id] = d;
+      return acc;
+    }, {});
 
-            return acc;
-          }, {}),
-        };
+    return { orderLinesMap };
+  }, []);
 
-        const newOrderLinesMap = {
-          ...orderLinesMap,
-          ...orderLinesResponse.reduce((acc, orderLine) => {
-            acc[orderLine.id] = {
-              ...orderLine,
-              locations: orderLine.locations.map(
-                ({ locationId }) => newLocationsMap[locationId]?.name ?? invalidReferenceMessage,
-              ),
-              orderWorkflow: newOrdersMap[orderLine.purchaseOrderId]?.workflowStatus,
-            };
-
-            return acc;
-          }, {}),
-        };
-
-        setOrderLinesMap(newOrderLinesMap);
-        setLocationsMap(newLocationsMap);
-        setOrdersMap(newOrdersMap);
-
-        setTitles((prev) => [
-          ...prev,
-          ...titlesResponse.titles.map(title => ({
-            ...title,
-            poLine: newOrderLinesMap[title.poLineId],
-          })),
-        ]);
-      });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [locationsMap, orderLinesMap, ordersMap]);
-
-  const {
-    records: titles,
-    recordsCount: titlesCount,
-    isLoading,
-    onNeedMoreData,
-  } = useList(false, loadTitles, loadTitlesCB, RESULT_COUNT_INCREMENT);
+  const { pagination, changePage } = usePagination({ limit: RESULT_COUNT_INCREMENT, offset: 0 });
+  const { titles, totalRecords, isFetching } = useReceiving({ pagination, fetchReferences });
 
   return (
     <ReceivingList
-      onNeedMoreData={onNeedMoreData}
+      onNeedMoreData={changePage}
       resetData={resetData}
-      titlesCount={titlesCount}
-      isLoading={isLoading}
+      titlesCount={totalRecords}
+      isLoading={isFetching}
       titles={titles}
+      pagination={pagination}
     />
   );
 };
@@ -142,7 +87,6 @@ ReceivingListContainer.manifest = Object.freeze({
 
 ReceivingListContainer.propTypes = {
   mutator: PropTypes.object.isRequired,
-  location: ReactRouterPropTypes.location.isRequired,
 };
 
-export default withRouter(stripesConnect(ReceivingListContainer));
+export default stripesConnect(ReceivingListContainer);