+++ src/ReceivingList/ReceivingListContainer.test.js
@@ -1,113 +1,84 @@
 import React from 'react';
-import { act, render, screen } from '@testing-library/react';
-import { MemoryRouter } from 'react-router';
-import queryString from 'query-string';
+import { render, screen, waitFor } from '@testing-library/react';
 
-import {
-  useList,
-} from '@folio/stripes-acq-components';
-
-import ReceivingList from './ReceivingList';
 import ReceivingListContainer from './ReceivingListContainer';
-import { buildTitlesQuery } from './utils';
+import { useReceiving } from './hooks';
 
 jest.mock('@folio/stripes-acq-components', () => ({
   ...jest.requireActual('@folio/stripes-acq-components'),
-  useList: jest.fn().mockReturnValue({}),
+  usePagination: () => ({}),
 }));
 jest.mock('./ReceivingList', () => jest.fn().mockReturnValue('ReceivingList'));
+jest.mock('./hooks/useReceiving', () => ({
+  useReceiving: jest.fn(),
+}));
 
 const defaultProps = {
   mutator: {
-    receivingListTitles: {
-      GET: jest.fn().mockReturnValue(Promise.resolve([])),
-    },
     receivingListOrderLines: {
-      GET: jest.fn().mockReturnValue(Promise.resolve([{ purchaseOrderId: 'orderId', locations: [{ locationId: 'id' }] }])),
+      GET: jest.fn().mockReturnValue(
+        Promise.resolve([{ purchaseOrderId: 'orderId', locations: [{ locationId: 'locationId' }] }]),
+      ),
     },
     lineOrders: {
       GET: jest.fn().mockReturnValue(Promise.resolve([{ id: 'orderId' }])),
     },
     receivingListLocations: {
-      GET: jest.fn().mockReturnValue(Promise.resolve([{ id: 'id' }])),
+      GET: jest.fn().mockReturnValue(Promise.resolve([{ id: 'locationId' }])),
     },
   },
-  location: {},
-  history: {},
 };
 
-const renderReceivingListContainer = (props = defaultProps) => render(
-  <ReceivingListContainer {...props} />,
-  { wrapper: MemoryRouter },
+const renderReceivingListContainer = (props = {}) => render(
+  <ReceivingListContainer
+    {...defaultProps}
+    {...props}
+  />,
 );
 
 describe('ReceivingListContainer', () => {
+  const titles = [{
+    id: '81f3f271-e6be-4ea1-93cd-edee96cc2227',
+    title: 'Multi-line titles #2',
+    poLineId: '3e1a947f-a605-41b8-839c-7929f02ef911',
+  }];
+
   beforeEach(() => {
-    defaultProps.mutator.receivingListTitles.GET.mockClear();
+    useReceiving
+      .mockClear()
+      .mockReturnValue({ titles, totalRecords: titles.length });
   });
 
-  it('should display Receiving list', async () => {
-    await act(async () => renderReceivingListContainer());
+  it('should display Receiving list', () => {
+    renderReceivingListContainer();
 
     expect(screen.getByText('ReceivingList')).toBeDefined();
   });
 
-  it('should pass useList result to ReceivingList', () => {
-    const records = [{ id: 'titleId' }];
-
-    ReceivingList.mockClear();
-    useList.mockClear().mockReturnValue({ records });
+  it('should load order lines, orders and receiving locations when fetchReferences is called', async () => {
     renderReceivingListContainer();
 
-    expect(ReceivingList.mock.calls[0][0].titles).toBe(records);
-  });
-
-  it('should load titles in useList', () => {
-    defaultProps.mutator.receivingListTitles.GET.mockClear();
-    useList.mockClear();
-
-    renderReceivingListContainer();
-
-    useList.mock.calls[0][1](5);
-
-    expect(defaultProps.mutator.receivingListTitles.GET).toHaveBeenCalledWith({
-      params: {
-        limit: 30,
-        offset: 5,
-        query: '(cql.allRecords=1) sortby title/sort.ascending',
-      },
-    });
-  });
-
-  it('should load title poLines in useList', async () => {
-    defaultProps.mutator.receivingListOrderLines.GET.mockClear();
-    useList.mockClear();
-
-    renderReceivingListContainer();
-
-    await act(async () => {
-      await useList.mock.calls[0][2](jest.fn(), { titles: [{ poLineId: 'id' }] });
-    });
+    await waitFor(() => useReceiving.mock.calls[0][0].fetchReferences(titles));
 
     expect(defaultProps.mutator.receivingListOrderLines.GET).toHaveBeenCalledWith({
       params: {
         limit: 1000,
-        query: 'id==id',
+        query: `id==${titles[0].poLineId}`,
       },
     });
-  });
 
-  describe('search query', () => {
-    it('should build query when search is active', () => {
-      const expectedQuery = '((title=="*query*" or poLine.titleOrPackage=="*query*" or productIds=="*query*" or purchaseOrder.poNumber=="*query*" or poLine.poLineNumber=="*query*" or poLine.vendorDetail.referenceNumbers=="*query*")) sortby title/sort.ascending';
-
-      expect(buildTitlesQuery(queryString.parse('?query=query'))).toBe(expectedQuery);
+    expect(defaultProps.mutator.lineOrders.GET).toHaveBeenCalledWith({
+      params: {
+        limit: 1000,
+        query: 'id==orderId',
+      },
     });
 
-    it('should build query when search by field is active', () => {
-      const expectedQuery = '(((title==*title*))) sortby title/sort.ascending';
-
-      expect(buildTitlesQuery(queryString.parse('?qindex=title&query=title'))).toBe(expectedQuery);
+    expect(defaultProps.mutator.receivingListLocations.GET).toHaveBeenCalledWith({
+      params: {
+        limit: 1000,
+        query: 'id==locationId',
+      },
     });
   });
 });